pyproject.toml
[build-system]
requires = [
    "setuptools >= 65",
    "wheel >= 0.38",
    "pyinstaller"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["ctksidebar"]

[project]
name = "ctk-sidebar"
version = "0.1.0"
authors = [
    { name="Anthony Bernaert" }
]
description = "A modern sidebar navigation control for CustomTkinter"
dependencies = [
    'pillow',
    'customtkinter'
]
requires-python = ">=3.9"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only"
]
license = "MIT"
license-files = ["LICENSE*"]

[project.urls]
Homepage = "https://github.com/anthony-bernaert/ctk-sidebar"
Repository = "https://github.com/anthony-bernaert/ctk-sidebar"
Issues = "https://github.com/anthony-bernaert/ctk-sidebar/issues"